+++ pidgin/gtkthemes.c
@@ -315,14 +315,14 @@ void pidgin_themes_load_smiley_theme(const char *file, gboolean load)
 			}
 			while  (*i) {
 				char l[64];
-				int li = 0;
+				size_t li = 0;
 				char *next;
 				while (*i && !isspace(*i) && li < sizeof(l) - 1) {
 					if (*i == '\\' && *(i+1) != '\0')
 						i++;
 					/* coverity[tainted_data] */
 					next = g_utf8_next_char(i);
-					if ((next - i) > (sizeof(l) - li -1)) {
+					if ((size_t)(next - i) > (sizeof(l) - li -1)) {
 						break;
 					}
 					while (i != next)
